Fixed types emit
